Add global exception handler for managing run-time exceptions
When a exception occurs in controllers or services,
It can be handlered in a global exception handler which is annotated with @RestControllerAdvice.

That is, all the exceptions are gathered into the handler class.

In the handler class, each method is assigned to each exception
and the method returns Response Entity which is a wrapper for ErrorResponse.

All error codes are managed in a ErrorCode implemented by enum.

So when you handle the exceptions, you can follow the process below.

1) Make your own exception class extended by RuntimeException.

2) Make a method to handle the exception in the global exception handler.

3) Make a corresponding error code enum type in the ErrorCode.